refactor: split export routes into modular directory structure
Decompose 827-line server/src/routes/export.ts into a directory mirroring
the bins/locations route patterns:

  routes/export/
    index.ts   composition + auth
    json.ts    GET /export, POST /import (JSON streaming)
    csv.ts     GET /export/csv, POST /import/csv
    zip.ts    GET /export/zip, POST /import/zip

Hoist split-revealed duplicates into the existing lib layer rather than
leaving them in two route files:
  - MAX_EXPORT_BINS / MAX_EXPORT_PHOTOS_PER_BIN / MAX_EXPORT_TOTAL_PHOTOS
    -> lib/exportHelpers.ts
  - csvEscape (was nested in handler closure) and parseCsvTags
    -> lib/csvParser.ts
  - logImportActivity wrapper (was triplicated across handlers)
    -> lib/routeHelpers.ts

Mounting unchanged: ESM resolves './routes/export/index.js' from index.ts.
All 35 export.test.ts + importSecurity.test.ts cases remain green; full
server suite (1545 tests) passes.
