+++ server/src/lib/csvParser.ts
@@ -55,3 +55,29 @@ export function parseCSV(text: string): string[][] {
 
   return rows;
 }
+
+/**
+ * Escape a value for CSV output. Guards against formula injection (Excel/Sheets
+ * interpret =, +, -, @, \t, \r as formulas) by prefixing the value with a single
+ * quote when needed, and quotes/escapes any value containing a delimiter or quote.
+ */
+export function csvEscape(val: string): string {
+  if (/^[=+\-@\t\r]/.test(val)) {
+    val = `'${val}`;
+  }
+  if (val.includes(',') || val.includes('"') || val.includes('\n') || val.includes('\r') || val.includes("'")) {
+    return `"${val.replace(/"/g, '""')}"`;
+  }
+  return val;
+}
+
+/**
+ * Parse a tags field from CSV.
+ * New exports use "; " (semicolon-space) as the delimiter to avoid splitting
+ * tags that contain commas (e.g. "nuts, bolts"). Legacy exports used bare
+ * commas, so we fall back to comma-splitting when no semicolons are present.
+ */
+export function parseCsvTags(raw: string): string[] {
+  const sep = raw.includes(';') ? ';' : ',';
+  return raw.split(sep).map(t => t.trim()).filter(Boolean);
+}
